requirements.txt
langchain~=0.3.27
langchain_core
langchain_community
langchain-google_genai
sqlmodel~=0.0.24
sqlalchemy~=2.0.43
fastapi
uvicorn
chromadb
langchain-chroma
python-dotenv
requests
pydantic
aiohttp
anyio~=4.10.0
langchain-community~=0.3.29
greenlet
aiosqlite
sentence-transformers
langchain-huggingface
streamlit
